Refactor sandbox state handling in worker manager
